refactor lemma utility class
