+++ src/main/resources/application.yaml
@@ -30,7 +30,7 @@ spring:
     database: mysql
     show-sql: false
     hibernate:
-      ddl-auto: update
+      ddl-auto: create-drop
       dialect: org.hibernate.dialect.MySQL8Dialect
       naming:
         physical-strategy: org.hibernate.boot.model.naming.PhysicalNamingStrategyStandardImpl
